Refactor: DooingleCount 디렉토리 분리
- UserService에서 DooingleCountService 의존하도록 수정
