+++ src/main/kotlin/com/dooingle/domain/dooinglecount/repository/DooingleCountRepository.kt
@@ -1,6 +1,6 @@
-package com.dooingle.domain.user.repository
+package com.dooingle.domain.dooinglecount.repository
 
-import com.dooingle.domain.user.model.DooingleCount
+import com.dooingle.domain.dooinglecount.model.DooingleCount
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface DooingleCountRepository : JpaRepository<DooingleCount, Long>, DooingleCountQueryDslRepository {
