+++ src/main/kotlin/com/dooingle/domain/dooinglecount/service/DooingleCountService.kt
@@ -1,6 +1,7 @@
-package com.dooingle.domain.user.service
+package com.dooingle.domain.dooinglecount.service
 
-import com.dooingle.domain.user.repository.DooingleCountRepository
+import com.dooingle.domain.user.dto.DooinglerResponse
+import com.dooingle.domain.dooinglecount.repository.DooingleCountRepository
 import org.springframework.scheduling.annotation.Scheduled
 import org.springframework.stereotype.Service
 
@@ -8,6 +9,10 @@ import org.springframework.stereotype.Service
 class DooingleCountService(
     private val dooingleCountRepository: DooingleCountRepository
 ) {
+
+    fun getHotDooinglerList(): List<DooinglerResponse> {
+        return dooingleCountRepository.getHighCountDooinglers()
+    }
 
     // 매일 0시 0분 0초에 dooingle_count 모든 데이터 삭제
     @Scheduled(cron = "0 0 0 * * *", zone = "Asia/Seoul")
